! finish theme refactor ! consolidated all functions to the abstract, leaving onlyt the required getSettings() in the actual theme dir.  Nex broke down the monolithic theme.php abstract based on traits.
