+++ sources/ElkArte/Themes/Theme.php
@@ -13,24 +13,22 @@
 
 namespace ElkArte\Themes;
 
-use BBC\ParserWrapper;
-use ElkArte\Controller\ScheduledTasks;
-use ElkArte\EventManager;
-use ElkArte\Helper\FileFunctions;
 use ElkArte\Helper\HttpReq;
-use ElkArte\Helper\SiteCombiner;
-use ElkArte\Helper\Util;
 use ElkArte\Helper\ValuesContainer;
-use ElkArte\Http\Headers;
-use ElkArte\Languages\Txt;
-use ElkArte\Menu\MenuContext;
-use ElkArte\User;
+use ElkArte\Themes\Traits\ContextManagement;
+use ElkArte\Themes\Traits\FeatureIntegration;
+use ElkArte\Themes\Traits\TemplateRendering;
 
 /**
  * Class Theme
  */
 abstract class Theme
 {
+	// Abstract Class Traits
+	use TemplateRendering;
+	use ContextManagement;
+	use FeatureIntegration;
+
 	/** @var string */
 	public const DEFAULT_EXPIRES = 'Mon, 26 Jul 1997 05:00:00 GMT';
 
@@ -76,6 +74,12 @@ abstract class Theme
 	/** @var Css */
 	public $css;
 
+	/** @var AssetManager */
+	private $assetManager;
+
+	/** @var HeadersManager */
+	private $headersManager;
+
 	/**
 	 * Theme constructor.
 	 *
@@ -85,6 +89,8 @@ abstract class Theme
 	 */
 	public function __construct(int $id, ValuesContainer $user, Directories $dirs)
 	{
+		global $settings;
+
 		$this->id = $id;
 		$this->user = $user;
 		$this->layers = new TemplateLayers();
@@ -110,6 +116,10 @@ abstract class Theme
 		// Theme posse
 		$this->javascript = new Javascript();
 		$this->css = new Css();
+
+		// Initialize helper classes
+		$this->assetManager = new AssetManager($this->javascript, $this->css);
+		$this->headersManager = new HeadersManager($this->_req);
 	}
 
 	/**
@@ -117,208 +127,6 @@ abstract class Theme
 	 */
 	abstract public function getSettings();
 
-	/**
-	 * The header template
-	 *
-	 * What it does:
-	 *  - Performs security checks
-	 *  - Sets up the theme context
-	 *  - Sets up headers (expiration, content type)
-	 *  - Loads template layers
-	 *  - Sends headers
-	 */
-	public function template_header(): void
-	{
-		doSecurityChecks();
-
-		$this->setupThemeContext();
-
-		$header = Headers::instance();
-		$this->setupHeadersExpiration($header);
-		$this->setupHeadersContentType($header, $this->getRequestAPI());
-
-		foreach ($this->getLayers()->prepareContext() as $layer)
-		{
-			$this->getTemplates()->loadSubTemplate($layer . '_above', 'ignore');
-		}
-
-		$this->loadDefaultThemeSettings();
-
-		$header->sendHeaders();
-	}
-
-	/**
-	 * Sets up the basic theme context.
-	 *
-	 * What it does:
-	 *  - Sets the current time and action
-	 *  - Checks if the current action should be indexed by robots
-	 *  - Prepares search engines dropdown
-	 *  - Sets up news lines, user context, menu, PM popup, common stats, theme data
-	 *  - Loads custom CSS
-	 *
-	 * @param bool $forceload = false
-	 */
-	public function setupThemeContext($forceload = false): void
-	{
-		global $context;
-
-		static $loaded = false;
-
-		// Under SSI this function can be called more than once.  That can cause some problems.
-		// So only run the function once unless we are forced to run it again.
-		if ($loaded && !$forceload)
-		{
-			return;
-		}
-
-		$loaded = true;
-
-		$context['current_time'] = standardTime(time(), false);
-		$context['current_action'] = $this->_req->getQuery('action', 'trim', '');
-		$context['robot_no_index'] = in_array($context['current_action'], $this->no_index_actions, true);
-		$context['additional_dropdown_search'] = prepareSearchEngines();
-
-		$this->setupNewsLines();
-		$this->setupCurrentUserContext();
-		(new MenuContext())->setupMenuContext();
-		$this->setContextShowPmPopup();
-		$this->setContextCommonStats();
-		$this->setContextThemeData();
-		$this->loadCustomCSS();
-	}
-
-	/**
-	 * Sets up the information context for the current user
-	 *
-	 * What it does:
-	 *  - Sets the current time and current action
-	 *  - Checks if the current action should be indexed by robots
-	 *  - Calls setupLoggedUserContext if the user is not a guest
-	 *  - Calls setupGuestContext if the user is a guest
-	 *  - Checks if the PM popup should be shown and adds the necessary JavaScript code
-	 */
-	public function setupCurrentUserContext(): void
-	{
-		global $scripturl, $context, $options, $txt;
-
-		$context['current_time'] = standardTime(time(), false);
-		$context['current_action'] = $this->_req->getQuery('action', 'trim', '');
-		$context['robot_no_index'] = in_array($context['current_action'], $this->no_index_actions, true);
-
-		if ($this->user->is_guest === false)
-		{
-			$this->setupLoggedUserContext();
-		}
-		else
-		{
-			$this->setupGuestContext();
-		}
-
-		$context['show_pm_popup'] = $context['user']['popup_messages'] && !empty($options['popup_messages']) && $context['current_action'] !== 'pm';
-		if ($context['show_pm_popup'])
-		{
-			$this->addInlineJavascript('
-		$(function() {
-			new elk_Popup({
-				heading: ' . JavaScriptEscape($txt['show_personal_messages_heading']) . ',
-				content: ' . JavaScriptEscape(sprintf($txt['show_personal_messages'], $context['user']['unread_messages'], $scripturl . '?action=pm')) . ',
-				icon: \'i-envelope\'
-			});
-		});', true);
-		}
-	}
-
-	/**
-	 * Load custom CSS files and add CSS rules
-	 *
-	 * What it does:
-	 *  - Loads custom.css if it exists for the theme
-	 *  - Adds avatar resize rules
-	 *  - Adds forum wrapper width (can use important to override in theme css)
-	 *  - Sets show more quote rules (localization & --quote_height)
-	 *  - Sets the profile button avatar
-	 */
-	public function loadCustomCSS(): void
-	{
-		global $settings, $modSettings, $txt;
-
-		// Load a base theme custom CSS file?
-		$fileFunc = FileFunctions::instance();
-		if ($fileFunc->fileExists($settings['theme_dir'] . '/css/custom.css'))
-		{
-			loadCSSFile('custom.css');
-		}
-
-		// Since it's nice to have avatars all the same size, and in some cases the size detection may fail,
-		// let's add the css in any case
-		if (!empty($modSettings['avatar_max_width']) || !empty($modSettings['avatar_max_height']))
-		{
-			$this->css->addCSSRules('
-		.avatarresize {' . (empty($modSettings['avatar_max_width']) ? '' : '
-			max-width:' . $modSettings['avatar_max_width'] . 'px;') . (empty($modSettings['avatar_max_height']) ? '' : '
-			max-height:' . $modSettings['avatar_max_height'] . 'px;') . '
-		}');
-		}
-
-		// Save some database hits, if a width for multiple wrappers is set in admin.
-		if (!empty($settings['forum_width']))
-		{
-			$this->css->addCSSRules('
-		.wrapper {width: ' . $settings['forum_width'] . ';}');
-		}
-
-		// Localization for the show more quote and it's container height
-		$quote_height = empty($modSettings['heightBeforeShowMore']) ? 'none' : $modSettings['heightBeforeShowMore'] . 'px';
-		$this->css->addCSSRules('
-		input[type=checkbox].quote-show-more:after {content: "' . $txt['quote_expand'] . '";}
-		.quote-read-more > .bbc_quote {--quote_height: ' . $quote_height . ';}'
-		);
-
-		if (!empty($this->user->avatar['href']))
-		{
-			$this->css->addCSSRules('
-		.i-menu-profile::before, .i-menu-profile.enabled::before {
-			content: "";
-			background-image: url("' . htmlspecialchars_decode($this->user->avatar['href']) . '");
-			background-position: center;
-			filter: unset;
-		}');
-		}
-	}
-
-	/**
-	 * The template footer
-	 *
-	 * What it does:
-	 *  - Sets up load time display if enabled
-	 *  - Restores theme settings if using default images
-	 *  - Loads template layers in reverse order
-	 */
-	public function template_footer(): void
-	{
-		global $context, $settings, $modSettings, $time_start;
-
-		$db = database();
-
-		// Show the load time?  (only makes sense for the footer.)
-		$context['show_load_time'] = !empty($modSettings['timeLoadPageEnable']);
-		$context['load_time'] = round(microtime(true) - $time_start, 3);
-		$context['load_queries'] = $db->num_queries();
-
-		if (isset($settings['use_default_images'], $settings['default_template'])
-			&& $settings['use_default_images'] === 'defaults')
-		{
-			$settings['theme_url'] = $settings['actual_theme_url'];
-			$settings['images_url'] = $settings['actual_images_url'];
-			$settings['theme_dir'] = $settings['actual_theme_dir'];
-		}
-
-		foreach ($this->getLayers()->reverseLayers() as $layer)
-		{
-			$this->getTemplates()->loadSubTemplate($layer . '_below', 'ignore');
-		}
-	}
 
 	/**
 	 * Load the base JS that gives ElkArte functionality
@@ -331,32 +139,13 @@ abstract class Theme
 	 */
 	public function loadThemeJavascript(): void
 	{
-		global $settings, $context, $modSettings, $scripturl, $txt, $options;
+		global $modSettings;
 
-		// Queue our Javascript
-		loadJavascriptFile(['script.js', 'script_elk.js', 'elk_menu.js']);
-		loadJavascriptFile(['theme.js'], ['defer' => true]);
-
-		// Default JS variables for use in every theme
-		$this->addJavascriptVar([
-			'elk_theme_url' => JavaScriptEscape($settings['theme_url']),
-			'elk_default_theme_url' => JavaScriptEscape($settings['default_theme_url']),
-			'elk_images_url' => JavaScriptEscape($settings['images_url']),
-			'elk_smiley_url' => JavaScriptEscape($modSettings['smileys_url']),
-			'elk_scripturl' => "'" . $scripturl . "'",
-			'elk_charset' => '"UTF-8"',
-			'elk_session_id' => JavaScriptEscape($context['session_id']),
-			'elk_session_var' => JavaScriptEscape($context['session_var']),
-			'elk_member_id' => $context['user']['id'],
-			'ajax_notification_text' => JavaScriptEscape($txt['ajax_in_progress']),
-			'ajax_notification_cancel_text' => JavaScriptEscape($txt['modify_cancel']),
-			'help_popup_heading_text' => JavaScriptEscape($txt['help_popup']),
-			'use_click_menu' => empty($options['use_click_menu']) ? 'false' : 'true',
-			'todayMod' => empty($modSettings['todayMod']) ? 0 : (int) $modSettings['todayMod']]
-		);
+		// Delegate base JavaScript loading to AssetManager
+		$this->assetManager->loadThemeJavascript();
 
 		// PWA?
-		$this->progressiveWebApp();
+		$this->assetManager->progressiveWebApp();
 
 		// Auto video embedding enabled, then load the needed JS
 		$this->autoEmbedVideo();
@@ -418,219 +207,7 @@ abstract class Theme
 	 */
 	public function getRequestAPI(): string|false
 	{
-		$api = $this->_req->getRequest('api', 'trim', '');
-
-		return in_array($api, ['xml', 'json', 'html']) && !empty($_SERVER['HTTP_X_REQUESTED_WITH']) ? $api : false;
-	}
-
-	/**
-	 * Set the headers expiration
-	 *
-	 * What it does:
-	 *  - Sets the Expires and Last-Modified headers in the Headers object.
-	 *
-	 * @param Headers $header The Headers object to set the headers in.
-	 */
-	public function setupHeadersExpiration(Headers $header): void
-	{
-		global $context;
-
-		if (empty($context['no_last_modified']))
-		{
-			$header
-				->header('Expires', self::DEFAULT_EXPIRES)
-				->header('Last-Modified', gmdate('D, d M Y H:i:s') . ' GMT');
-		}
-	}
-
-	/**
-	 * Set up the logged user context
-	 *
-	 * What it does:
-	 *  - Copies relevant user data from the user object to the global context.
-	 */
-	public function setupLoggedUserContext(): void
-	{
-		global $context;
-
-		$context['user']['messages'] = $this->user->messages;
-		$context['user']['unread_messages'] = $this->user->unread_messages;
-		$context['user']['mentions'] = $this->user->mentions;
-
-		// Personal message popup...
-		$context['user']['popup_messages'] = $this->user->unread_messages > ($_SESSION['unread_messages'] ?? 0);
-
-		$_SESSION['unread_messages'] = $this->user->unread_messages;
-
-		$context['user']['avatar'] = [
-			'href' => empty($this->user->avatar['href']) ? '' : $this->user->avatar['href'],
-			'image' => empty($this->user->avatar['image']) ? '' : $this->user->avatar['image'],
-		];
-
-		// Figure out how long they've been logged in.
-		$context['user']['total_time_logged_in'] = [
-			'days' => floor($this->user->total_time_logged_in / 86400),
-			'hours' => floor(($this->user->total_time_logged_in % 86400) / 3600),
-			'minutes' => floor(($this->user->total_time_logged_in % 3600) / 60)
-		];
-	}
-
-	/**
-	 * Set up guest context
-	 *
-	 * What it does:
-	 *  - Initializes global variables for guest user context
-	 */
-	public function setupGuestContext(): void
-	{
-		global $modSettings, $context, $txt;
-
-		$context['user']['messages'] = 0;
-		$context['user']['unread_messages'] = 0;
-		$context['user']['mentions'] = 0;
-		$context['user']['avatar'] = [];
-		$context['user']['total_time_logged_in'] = ['days' => 0, 'hours' => 0, 'minutes' => 0];
-		$context['user']['popup_messages'] = false;
-
-		if (!empty($modSettings['registration_method']) && (int) $modSettings['registration_method'] === 1)
-		{
-			$txt['welcome_guest'] .= $txt['welcome_guest_activate'];
-		}
-
-		$txt['welcome_guest'] = replaceBasicActionUrl($txt['welcome_guest']);
-	}
-
-	/**
-	 * Set the common stats in the context
-	 *
-	 * What it does:
-	 *  - Sets the total posts, total topics, total members, and latest member stats in the common_stats array of the context
-	 *  - Sets the formatted string for displaying the total posts in the boardindex_total_posts variable of the context
-	 */
-	public function setContextCommonStats(): void
-	{
-		global $context, $txt, $modSettings;
-
-		// This looks weird, but it's because BoardIndex.controller.php references the variable.
-		$href = getUrl('profile', ['action' => 'profile', 'u' => $modSettings['latestMember'], 'name' => $modSettings['latestRealName']]);
-
-		$context['common_stats'] = [
-			'total_posts' => comma_format($modSettings['totalMessages']),
-			'total_topics' => comma_format($modSettings['totalTopics']),
-			'total_members' => comma_format($modSettings['totalMembers']),
-			'latest_member' => [
-				'id' => $modSettings['latestMember'],
-				'name' => $modSettings['latestRealName'],
-				'href' => $href,
-				'link' => '<a href="' . $href . '">' . $modSettings['latestRealName'] . '</a>',
-			],
-		];
-
-		$context['common_stats']['boardindex_total_posts'] = sprintf($txt['boardindex_total_posts'], $context['common_stats']['total_posts'], $context['common_stats']['total_topics'], $context['common_stats']['total_members']);
-	}
-
-	/**
-	 * This is the only template included in the sources.
-	 */
-	public function template_rawdata(): void
-	{
-		global $context;
-
-		echo $context['raw_data'];
-	}
-
-	/**
-	 * Set the headers content type
-	 *
-	 * What it does:
-	 *  - Sets the content type of the headers based on the provided context and API.
-	 *
-	 * @param Headers $header The Headers instance used to set the content type.
-	 * @param string $api The API string used to determine the content type.
-	 */
-	public function setupHeadersContentType(Headers $header, string $api): void
-	{
-		$contentType = self::CONTENT_TYPES[$api] ?? 'text/html';
-
-		$header->contentType($contentType, 'UTF-8');
-	}
-
-	/**
-	 * Load default theme settings
-	 *
-	 * Updates the theme settings by replacing the URL and directory values with the default ones if the 'use_default_images'
-	 * setting is set to 'defaults' and the 'default_template' setting is provided.
-	 */
-	public function loadDefaultThemeSettings(): void
-	{
-		global $settings;
-
-		if (isset($settings['use_default_images'], $settings['default_template'])
-			&& $settings['use_default_images'] === 'defaults')
-		{
-			$settings['theme_url'] = $settings['default_theme_url'];
-			$settings['images_url'] = $settings['default_images_url'];
-			$settings['theme_dir'] = $settings['default_theme_dir'];
-		}
-	}
-
-	/**
-	 * Sets up the news lines for display
-	 *
-	 * What it does:
-	 *  - Retrieves the news lines from the modSettings variable
-	 *  - Filters out empty lines and trims whitespace
-	 *  - Parses the news lines using the BBC parser
-	 *  - Sets a random news line as the 'random_news_line' variable in the context
-	 *  - Adds the 'news_fader' callback to the 'upper_content_callbacks' array in the context
-	 *  - Sets the 'show_news' variable in the context based on the 'enable_news' setting in $settings
-	 */
-	public function setupNewsLines(): void
-	{
-		global $context, $modSettings, $settings;
-
-		$context['news_lines'] = array_filter(explode("\n", str_replace("\r", '', trim(addslashes($modSettings['news'])))));
-		$bbc_parser = ParserWrapper::instance();
-		foreach ($context['news_lines'] as $i => $iValue)
-		{
-			if (trim($iValue) === '')
-			{
-				continue;
-			}
-
-			$context['news_lines'][$i] = $bbc_parser->parseNews(stripslashes(trim($iValue)));
-		}
-
-		if (empty($context['news_lines']))
-		{
-			return;
-		}
-
-		$context['random_news_line'] = $context['news_lines'][mt_rand(0, count($context['news_lines']) - 1)];
-		$context['upper_content_callbacks'][] = 'news_fader';
-
-		// This is here because old index templates might still use it.
-		$context['show_news'] = !empty($settings['enable_news']);
-	}
-
-	/**
-	 * Show the copyright.
-	 */
-	public function theme_copyright(): void
-	{
-		global $forum_copyright;
-
-		// Don't display copyright for things like SSI.
-		if (!defined('FORUM_VERSION'))
-		{
-			return;
-		}
-
-		// Put in the version...
-		$forum_copyright = replaceBasicActionUrl(sprintf($forum_copyright, FORUM_VERSION));
-
-		echo '
-					', $forum_copyright;
+		return $this->headersManager->getRequestAPI();
 	}
 
 	/**
@@ -663,153 +240,7 @@ abstract class Theme
 	 */
 	public function cleanHives($type = 'all'): bool
 	{
-		global $settings;
-
-		$combiner = new SiteCombiner($settings['default_theme_cache_dir'], $settings['default_theme_cache_url']);
-		$result = true;
-
-		if ($type === 'all' || $type === 'css')
-		{
-			$result = $combiner->removeCssHives();
-		}
-
-		if ($type === 'all' || $type === 'js')
-		{
-			$result = $result && $combiner->removeJsHives();
-		}
-
-		// Force a cache refresh for the PWA
-		setPWACacheStale(true);
-
-		return $result;
-	}
-
-	/**
-	 * If video embedding is enabled, this loads the necessary JS and vars
-	 */
-	public function autoEmbedVideo(): void
-	{
-		global $txt, $modSettings;
-
-		if (!empty($modSettings['enableVideoEmbeding']))
-		{
-			loadJavascriptFile('elk_jquery_embed.js', ['defer' => true]);
-
-			$this->addInlineJavascript('
-				if (typeof oEmbedtext === "undefined") {
-					var oEmbedtext = ({
-						embed_limit : ' . (empty($modSettings['video_embed_limit']) ? 25 : $modSettings['video_embed_limit']) . ',
-						preview_image : ' . JavaScriptEscape($txt['preview_image']) . ',
-						ctp_video : ' . JavaScriptEscape($txt['ctp_video']) . ',
-						hide_video : ' . JavaScriptEscape($txt['hide_video']) . ',
-						youtube : ' . JavaScriptEscape($txt['youtube']) . ',
-						vimeo : ' . JavaScriptEscape($txt['vimeo']) . ',
-						dailymotion : ' . JavaScriptEscape($txt['dailymotion']) . ',
-						tiktok : ' . JavaScriptEscape($txt['tiktok']) . ',
-						twitter : ' . JavaScriptEscape($txt['twitter']) . ',
-						facebook : ' . JavaScriptEscape($txt['facebook']) . ',
-						instagram : ' . JavaScriptEscape($txt['instagram']) . ',
-					});
-
-					document.addEventListener("DOMContentLoaded", () => {
-						if ($.isFunction($.fn.linkifyvideo))
-						{
-							$().linkifyvideo(oEmbedtext);
-						}
-					});
-				}
-			', true);
-		}
-	}
-
-	/**
-	 * Progressive Web App initialization
-	 *
-	 * What it does:
-	 *  - Sets up the necessary configurations for the Progressive Web App (PWA).
-	 *  - Adds JavaScript variables, loads necessary JavaScript files, and adds inline JavaScript code.
-	 *
-	 * @return void
-	 */
-	public function progressiveWebApp(): void
-	{
-		global $modSettings, $boardurl, $settings;
-
-		$this->addJavascriptVar([
-			'elk_board_url' => JavaScriptEscape($boardurl),
-		]);
-		loadJavascriptFile('elk_pwa.js', ['defer' => false]);
-
-		// Not enabled, let's be sure to remove it should it exist
-		if (empty($modSettings['pwa_enabled']))
-		{
-			$this->addInlineJavascript('
-				elkPwa().removeServiceWorker();
-			');
-
-			return;
-		}
-
-		setPWACacheStale();
-		$theme_scope = $this->getScopeFromUrl($settings['actual_theme_url']);
-		$default_theme_scope = $this->getScopeFromUrl($settings['default_theme_url']);
-		$sw_scope = $this->getScopeFromUrl($boardurl);
-		$this->addInlineJavascript('
-			document.addEventListener("DOMContentLoaded", function() {
-				let myOptions = {
-					swUrl: "elkServiceWorker.js",
-					swOpt: {
-						cache_stale: ' . JavaScriptEscape(CACHE_STALE) . ',
-						cache_id: ' . JavaScriptEscape($modSettings['elk_pwa_cache_stale']) . ',
-						theme_scope: ' . JavaScriptEscape($theme_scope) . ',
-						default_theme_scope: ' . JavaScriptEscape($default_theme_scope) . ',
-						sw_scope: ' . JavaScriptEscape($sw_scope) . ',
-						nav_preload: 1, // set to 1 to enable, 0 to disable
-					}
-				};
-	
-				let elkPwaInstance = elkPwa(myOptions);
-				elkPwaInstance.init();
-				elkPwaInstance.sendMessage("deleteOldCache", {cache_id: ' . JavaScriptEscape($modSettings['elk_pwa_cache_stale']) . '});
-				elkPwaInstance.sendMessage("pruneCache");
-			});'
-		);
-	}
-
-	/**
-	 * Get the scope from the given URL
-	 *
-	 * @param string $url The URL from which to extract the scope
-	 *
-	 * @return string The scope extracted from the URL, or the root scope if not found
-	 */
-	public function getScopeFromUrl($url): string
-	{
-		$parts = parse_url($url);
-
-		return empty($parts['path']) ? '/' : '/' . trim($parts['path'], '/') . '/';
-	}
-
-	/**
-	 * If the option to pretty output code is on, this loads the JS and CSS
-	 */
-	public function addCodePrettify(): void
-	{
-		global $modSettings;
-
-		if (!empty($modSettings['enableCodePrettify']))
-		{
-			$this->loadVariant('prettify');
-			loadJavascriptFile('ext/prettify.min.js', ['defer' => true]);
-
-			$this->addInlineJavascript('
-				document.addEventListener("DOMContentLoaded", () => {
-				if (typeof prettyPrint === "function")
-				{
-					prettyPrint();
-				}
-			});', true);
-		}
+		return $this->assetManager->cleanHives($type);
 	}
 
 	/**
@@ -821,307 +252,7 @@ abstract class Theme
 	 */
 	public function loadVariant($cssFile, $fallBack = true): void
 	{
-		global $settings, $context;
-
-		$fileFunc = FileFunctions::instance();
-		if ($fileFunc->fileExists($settings['theme_dir'] . '/css/' . $context['theme_variant'] . '/' . $cssFile . $context['theme_variant'] . '.css'))
-		{
-			loadCSSFile($context['theme_variant'] . '/' . $cssFile . $context['theme_variant'] . '.css');
-			return;
-		}
-
-		if (!$fallBack)
-		{
-			return;
-		}
-
-		if (!$fileFunc->fileExists($settings['theme_dir'] . '/css/' . $cssFile . '.css'))
-		{
-			return;
-		}
-
-		loadCSSFile($cssFile . '.css');
-	}
-
-	/**
-	 * Relative times require a few variables be set in the JS
-	 */
-	public function relativeTimes(): void
-	{
-		global $modSettings, $context, $txt;
-
-		// Relative times?
-		if (!empty($modSettings['todayMod']) && $modSettings['todayMod'] > 2)
-		{
-			loadJavascriptFile('elk_relativeTime.js', ['defer' => true]);
-			$this->addInlineJavascript('
-				if (typeof oRttime === "undefined") {
-					var oRttime = ({
-						referenceTime : ' . forum_time() * 1000 . ',
-						now : ' . JavaScriptEscape($txt['rt_now']) . ',
-						minute : ' . JavaScriptEscape($txt['rt_minute']) . ',
-						minutes : ' . JavaScriptEscape($txt['rt_minutes']) . ',
-						hour : ' . JavaScriptEscape($txt['rt_hour']) . ',
-						hours : ' . JavaScriptEscape($txt['rt_hours']) . ',
-						day : ' . JavaScriptEscape($txt['rt_day']) . ',
-						days : ' . JavaScriptEscape($txt['rt_days']) . ',
-						week : ' . JavaScriptEscape($txt['rt_week']) . ',
-						weeks : ' . JavaScriptEscape($txt['rt_weeks']) . ',
-						month : ' . JavaScriptEscape($txt['rt_month']) . ',
-						months : ' . JavaScriptEscape($txt['rt_months']) . ',
-						year : ' . JavaScriptEscape($txt['rt_year']) . ',
-						years : ' . JavaScriptEscape($txt['rt_years']) . ',
-					});
-				}
-				document.addEventListener("DOMContentLoaded", () => {updateRelativeTime();});', true);
-
-			$context['using_relative_time'] = true;
-		}
-	}
-
-	/**
-	 * Ensures we kick the mail queue from time to time so that it gets
-	 * checked as often as possible.
-	 */
-	public function doScheduledSendMail(): void
-	{
-		global $modSettings;
-
-		if (!empty(User::$info->possibly_robot))
-		{
-			// @todo Maybe move this somewhere better?!
-			$controller = new ScheduledTasks(new EventManager());
-
-			// What to do, what to do?!
-			if (empty($modSettings['next_task_time']) || $modSettings['next_task_time'] < time())
-			{
-				$controller->action_autotask();
-			}
-			else
-			{
-				$controller->action_reducemailqueue();
-			}
-		}
-		else
-		{
-			$type = empty($modSettings['next_task_time']) || $modSettings['next_task_time'] < time() ? 'task' : 'mailq';
-			$ts = $type === 'mailq' ? $modSettings['mail_next_send'] : $modSettings['next_task_time'];
-
-			$this->addInlineJavascript('
-		function elkAutoTask()
-		{
-			let tempImage = new Image();
-			tempImage.src = elk_scripturl + "?scheduled=' . $type . ';ts=' . $ts . '";
-		}
-		window.setTimeout("elkAutoTask();", 1);', true);
-		}
-	}
-
-	/**
-	 * Set the context for showing the PM popup
-	 *
-	 * What it does:
-	 *  - Sets the context variable $context['show_pm_popup'] based on user preferences and current action
-	 */
-	public function setContextShowPmPopup(): void
-	{
-		global $context, $options, $txt, $scripturl;
-
-		// This is done to allow theme authors to customize it as they want.
-		$context['show_pm_popup'] = $context['user']['popup_messages'] && !empty($options['popup_messages']) && $context['current_action'] !== 'pm';
-
-		// Add the PM popup here instead. Theme authors can still override it simply by editing/removing the 'fPmPopup' in the array.
-		if ($context['show_pm_popup'])
-		{
-			$this->addInlineJavascript('
-		$(function() {
-			new elk_Popup({
-				heading: ' . JavaScriptEscape($txt['show_personal_messages_heading']) . ',
-				content: ' . JavaScriptEscape(sprintf($txt['show_personal_messages'], $context['user']['unread_messages'], $scripturl . '?action=pm')) . ',
-				icon: \'i-envelope\'
-			});
-		});', true);
-		}
-	}
-
-	/**
-	 * Set the context theme data
-	 *
-	 * What it does:
-	 *  - Sets the theme data in the context array
-	 *  - Adds necessary JavaScript variables
-	 *  - Sets the page title and favicon
-	 *  - Updates the HTML headers
-	 */
-	public function setContextThemeData(): void
-	{
-		global $context, $scripturl, $settings, $boardurl, $modSettings, $txt, $mbname;
-
-		if (empty($settings['theme_version']))
-		{
-			$this->addJavascriptVar(['elk_scripturl' => $scripturl], true);
-		}
-
-		$this->addJavascriptVar(['elk_forum_action' => getUrlQuery('action', $modSettings['default_forum_action'])], true);
-
-		$context['page_title'] = $context['page_title'] ?? $mbname;
-		$context['page_title_html_safe'] = Util::htmlspecialchars(un_htmlspecialchars($context['page_title'])) . (empty($context['current_page']) ? '' : ' - ' . $txt['page'] . (' ' . ($context['current_page'] + 1)));
-		$context['favicon'] = $boardurl . '/favicon.ico';
-		$context['apple_touch'] = $boardurl . '/themes/default/images/logos/apple-touch-icon.png';
-		$context['html_headers'] = $context['html_headers'] ?? '';
-		$context['theme-color'] = $modSettings['pwa_theme-color'] ?? '#3d6e32';
-		$context['pwa_manifest_enabled'] = !empty($modSettings['pwa_manifest_enabled']);
-	}
-
-	/**
-	 * If a variant CSS is needed, this loads it
-	 */
-	public function loadThemeVariant(): void
-	{
-		global $context, $settings, $options;
-
-		// Overriding - for previews and that ilk.
-		$variant = $this->_req->getRequest('variant', 'trim', '');
-		if (!empty($variant))
-		{
-			$_SESSION['id_variant'] = $variant;
-		}
-
-		// User selection?
-		if (empty($settings['disable_user_variant']) || allowedTo('admin_forum'))
-		{
-			$context['theme_variant'] = empty($_SESSION['id_variant']) ? (!empty($options['theme_variant']) ? $options['theme_variant'] : '') : ($_SESSION['id_variant']);
-		}
-
-		// If not a user variant, select the default.
-		if ($context['theme_variant'] === '' || !in_array($context['theme_variant'], $settings['theme_variants']))
-		{
-			$context['theme_variant'] = !empty($settings['default_variant']) && in_array($settings['default_variant'], $settings['theme_variants']) ? $settings['default_variant'] : $settings['theme_variants'][0];
-		}
-
-		// Do this to keep things easier in the templates.
-		$context['theme_variant'] = '_' . $context['theme_variant'];
-		$context['theme_variant_url'] = $context['theme_variant'] . '/';
-
-		// The most efficient way of writing multi themes is to use a master index.css plus variant.css files.
-		if (!empty($context['theme_variant']))
-		{
-			loadCSSFile($context['theme_variant'] . '/index' . $context['theme_variant'] . '.css');
-
-			// Variant icon definitions?
-			$this->loadVariant('icons_svg', false);
-
-			// Load a theme variant custom CSS
-			$this->loadVariant('custom', false);
-		}
-	}
-
-	/**
-	 * Calls on template_show_error from index.template.php to show warnings
-	 * and security errors for admins
-	 */
-	public function template_admin_warning_above(): void
-	{
-		global $context, $txt;
-
-		if (!empty($context['security_controls_files']))
-		{
-			$context['security_controls_files']['type'] = 'serious';
-			template_show_error('security_controls_files');
-		}
-
-		if (!empty($context['security_controls_query']))
-		{
-			$context['security_controls_query']['type'] = 'serious';
-			template_show_error('security_controls_query');
-		}
-
-		if (!empty($context['security_controls_ban']))
-		{
-			$context['security_controls_ban']['type'] = 'serious';
-			template_show_error('security_controls_ban');
-		}
-
-		if (!empty($context['new_version_updates']))
-		{
-			template_show_error('new_version_updates');
-		}
-
-		if (!empty($context['accepted_agreement']))
-		{
-			template_show_error('accepted_agreement');
-		}
-
-		// Any special notices to remind the admin about?
-		if (!empty($context['warning_controls']))
-		{
-			$context['warning_controls']['errors'] = $context['warning_controls'];
-			$context['warning_controls']['title'] = $txt['admin_warning_title'];
-			$context['warning_controls']['type'] = 'warning';
-			template_show_error('warning_controls');
-		}
-	}
-
-	/**
-	 * Makes the default layers and languages available
-	 *
-	 * - Loads index and addon language files as needed
-	 * - Loads XML, index, or no templates as needed
-	 * - Loads templates as defined by $settings['theme_templates']
-	 */
-	public function loadDefaultLayers(): void
-	{
-		global $settings;
-
-		$simpleActions = [
-			'quickhelp',
-			'printpage',
-			'quotefast',
-		];
-
-		call_integration_hook('integrate_simple_actions', [&$simpleActions]);
-
-		// Output is fully XML and sent by our JavaScript
-		$api = $this->_req->getRequest('api', 'trim', '');
-		$valid = !empty($_SERVER['HTTP_X_REQUESTED_WITH']);
-		$action = $this->_req->getRequest('action', 'trim', '');
-
-		if ($valid && $api === 'xml')
-		{
-			Txt::load('index+Addons');
-			$this->getLayers()->removeAll();
-			$this->getTemplates()->load('Xml');
-		}
-		// These actions don't require the index template at all.
-		elseif (in_array($action, $simpleActions, true))
-		{
-			Txt::load('index+Addons');
-			$this->getLayers()->removeAll();
-		}
-		else
-		{
-			// Custom templates to load, or just default?
-			$templates = isset($settings['theme_templates']) ? explode(',', $settings['theme_templates']) : ['index'];
-
-			// Load each template...
-			foreach ($templates as $template)
-			{
-				$this->getTemplates()->load($template);
-			}
-
-			// ...and attempt to load their associated language files.
-			Txt::load(array_merge($templates, ['Addons']), false);
-
-			// Custom template layers?
-			$layers = isset($settings['theme_layers']) ? explode(',', $settings['theme_layers']) : ['html', 'body'];
-
-			$template_layers = $this->getLayers();
-			$template_layers->setErrorSafeLayers($layers);
-			foreach ($layers as $layer)
-			{
-				$template_layers->addBegin($layer);
-			}
-		}
+		$this->assetManager->loadVariant($cssFile, $fallBack);
 	}
 
 	/**
@@ -1138,5 +269,13 @@ abstract class Theme
 	public function themeJs(): Javascript
 	{
 		return $this->javascript;
+	}
+
+	/**
+	 * Load theme variant CSS
+	 */
+	public function loadThemeVariant(): void
+	{
+		$this->assetManager->loadThemeVariant($this->_req);
 	}
 }
